requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-dateutil==2.8.2
jinja2==3.1.3
weasyprint==60.2
